refactor kubejobs plugin
this commit introduces a major refactor in how
kubejobs plugin acts. changes are listed below:

* introduce a generic approach that allows workflows
to be non-scone. now the manager does not deal with scone
cli at all, leaving all the login/split work to the user.
hence all the applications are the same from the manager point-of-view,
regardless of being scone/non-scone.

* volumes and environment variables are now populated by the manager
with information passed through the submission body. all volumes are
considered to be hostPath mounts (the kind we use for /dev/isgx mounts).

* remove any scone dependencies.

* controller and monitor are only started if their configuration is
defined in the submission body. they are not started otherwise.

* the manager still provisions a redis database for work queues, but the
name is previously defined by the user through ``redis_name`` element of
submission body. this is intended to be temporary.

there are still a few things that need to be discussed and improved. some of
them are listed below:

* improve provisioning of redis queues (find a way to stop defining hard-coded names).
* improve management of results (needs discussion).
* improve workload submission error flow (remove the logic from a giant try/except block).
* improve resource management (check if request is correct before doing work; make sure
things are deleted afterwards).